Detailed performance logging
The common class now has start and stop timer functions to allow for timing of various steps in the descendant class functions if needed. It will also email the log out. Also includes some basic memory functions. The StationConnection class now uses the timer functions to time the workflow of data acquisition from the weather station weather link live device to forwarding the data to the web app server. If it takes too long, an email alert is send out with the timings at each step along with memory usage.